+++ x-pack/filebeat/input/cel/config.go
@@ -114,18 +114,6 @@ func (c config) Validate() error {
 	return nil
 }
 
-// checkUnsupportedParams checks if unsupported/deprecated/discouraged paramaters are set and logs a warning
-func (c config) checkUnsupportedParams(logger *logp.Logger) {
-	if c.RecordCoverage {
-		logp.L().Named("cel").Warn("execution coverage enabled: " +
-			"see documentation for details: https://www.elastic.co/guide/en/beats/filebeat/current/filebeat-input-cel.html#cel-record-coverage")
-	}
-	if c.Redact == nil {
-		logp.L().Named("cel").Warn("missing recommended 'redact' configuration: " +
-			"see documentation for details: https://www.elastic.co/guide/en/beats/filebeat/current/filebeat-input-cel.html#cel-state-redact")
-	}
-}
-
 func defaultConfig() config {
 	maxExecutions := defaultMaxExecutions
 	maxAttempts := 5
